Completed set_to_pin method for digital pins
The method searches for the specified pin in a set of reserved pins.
If it does not find it, it means that the pin is not reserved and the method returns false.
If the Method finds the pin, but the specified value does not match the pin type, the method returns false.
Otherwise, the method returns true.